[nan-220] move internal-nango outside for readability and track runtime via a metric
