+++ packages/shared/lib/integrations/scripts/webhook/github-app-webhook-routing.ts
@@ -1,4 +1,4 @@
-import type { InternalNango as Nango } from './webhook.manager.js';
+import type { InternalNango as Nango } from './internal-nango.js';
 import type { Config as ProviderConfig } from '../../../models/Provider.js';
 import crypto from 'crypto';
