+++ packages/shared/lib/integrations/scripts/webhook/hubspot-webhook-routing.unit.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi } from 'vitest';
 import crypto from 'crypto';
 import * as HubspotWebhookRouting from './hubspot-webhook-routing.js';
-import type { InternalNango as Nango } from './webhook.manager.js';
+import type { InternalNango as Nango } from './internal-nango.js';
 import type { Config as ProviderConfig } from '../../../models/Provider.js';
 
 vi.mock('crypto', async () => {
