+++ packages/shared/lib/integrations/scripts/webhook/jira-webhook-routing.ts
@@ -1,4 +1,4 @@
-import type { InternalNango as Nango } from './webhook.manager.js';
+import type { InternalNango as Nango } from './internal-nango.js';
 import type { Config as ProviderConfig } from '../../../models/Provider.js';
 
 export default async function route(nango: Nango, integration: ProviderConfig, _headers: Record<string, any>, body: any) {
