+++ packages/shared/lib/integrations/scripts/webhook/slack-webhook-routing.ts
@@ -1,4 +1,5 @@
-import type { InternalNango as Nango, WebhookResponse } from './webhook.manager.js';
+import type { WebhookResponse } from './webhook.manager.js';
+import type { InternalNango as Nango } from './internal-nango.js';
 import type { Config as ProviderConfig } from '../../../models/Provider.js';
 
 export default async function route(
